SPY-108: Verify connections are sound before completing them entirely.
Without this change, the client considers the connection to be
good any time a connection has been established.  This is not the
case in some critical situations, such as a hung process or a very
busy remote server.

Solution is to ping the server with a version op, since that can happen
before auth, and use that as a way to verify the thing is actually
alive.

Note there is also a refactoring in here renaming the variable for
the node.

Change-Id: Ie3c0bc967b0705df6e58bf8ef81b158db8576bf3
Reviewed-on: http://review.couchbase.org/23840
Reviewed-by: Matt Ingenthron <matt@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com> (cherry-picking from commit 39d1a8dfca5f78e0a31bed96da669965833de7e4)