SPY-111: Check if the node belongs to cluster before connect.
This changeset makes sure a node belongs to the cluster (by
verifying against the node locator), before trying to (re)-
connect. This helps in situations where a node is scheduled
for shutdown and therfore not part of the cluster, but is
still alive and may be connectable. In this case, it will
result in NPE or AssertionExceptions down the stack.

Change-Id: Ib825b487fcd77c5cb9cf7d50e5b8b3c9c227cab5
Reviewed-on: http://review.couchbase.org/24230
Reviewed-by: Matt Ingenthron <matt@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com> (cherry-picking from commit 65cf4b068456e97306b88dbaaaf1811efaf9d008)