SPY-123: Fix Race Condition that kills IO Thread
This changeset makes sure that a potential race condition when
adding/removing nodes the same time does not kill the IO thread.
The Exception is caught and the code block is tried again.

Also, as a safety net, the IO thread at the topmost level catches
this error and proceeds, since this type of exception is considered
to be transient and does not persist.

Change-Id: I04bfea7f38327c96019f2fcccddcaccfe5b791d3
Reviewed-on: http://review.couchbase.org/26464
Reviewed-by: Deepti Dawar <deepti.dawar@globallogic.com>
Reviewed-by: Matt Ingenthron <matt@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com> (cherry-picking from commit 87c002a06d3fb084701bf79531906dc5dcb3df5d)