SPY-120: Adding support for SLF4J logging.
This changeset adds support for SLF4J logging. The Logger API
has been modified to contain only very small backwards breaking
changes, it only adds TRACE support to the interface. People
not implementing the Interface directly but just implementing
the AbstractLogger only need to add the new isTraceEnabled()
method and inside the log() method also check for the TRACE
level.

As SLF4J implies, this changeset does not ship with an
actual SLF4J implementation but lets you choose it at runtime
with just adding the impl to the classpath.

Change-Id: I178ef8c8dd7417a33d0e3914a644632b98c4a265
Reviewed-on: http://review.couchbase.org/26149
Reviewed-by: Matt Ingenthron <matt@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com> (cherry-picking from commit 730fe4f285887ed074c78da84ef58a85bf8abd4d)